+++ docs/main/changelog.rst
@@ -12,6 +12,7 @@ follows `Semantic versioning`_
 - Improve ``Configuration`` provider overriding logic.
 - Refactor ``Configuration`` provider.
 - Improve ``DependenciesContainer`` provider overriding logic.
+- Update "services" example miniapp.
 - Update "bundles" example miniapp.
 
 3.10.0